requirements.txt
pandas
faiss-cpu
torch
fuzzywuzzy
sentence-transformers
transformers
gradio